Convert Context API stores to Zustand
