+++ frontend/app/game.tsx
@@ -7,16 +7,23 @@ import { RootNavigator } from "./navigation/RootNavigator";
 import { useEventManager } from "./context/EventManager";
 import { usePowContractConnector } from "./context/PowContractConnector";
 import { useGame } from "./context/Game";
-import { useInAppNotifications } from "./context/InAppNotifications";
-import { useAchievement } from "./context/Achievements";
-import { useSound } from "./context/Sound";
+import { useInAppNotifications } from "./stores/useInAppNotificationsStore";
+import {
+  useAchievement,
+  useAchievementsStore,
+} from "./stores/useAchievementsStore";
+import { useSound, useSoundStore } from "./stores/useSoundStore";
 import { InAppNotificationsObserver } from "./observers/InAppNotificationsObserver";
 import { AchievementObserver } from "./observers/AchievementObserver";
 import { SoundObserver } from "./observers/SoundObserver";
 import { TxBuilderObserver } from "./observers/TxBuilderObserver";
+import { TutorialObserver } from "./observers/TutorialObserver";
+import { useTutorial, useTutorialStore } from "./stores/useTutorialStore";
 
 export default function game() {
   const { registerObserver, unregisterObserver } = useEventManager();
+  const { setSoundDependency, initializeAchievements } = useAchievementsStore();
+  const { initializeSound } = useSoundStore();
 
   const { sendInAppNotification } = useInAppNotifications();
   const [inAppNotificationObserver, setInAppNotificationObserver] = useState<
@@ -70,6 +77,30 @@ export default function game() {
     }
     setSoundObserver(registerObserver(new SoundObserver(playSoundEffect)));
   }, [playSoundEffect]);
+
+  useEffect(() => {
+    initializeSound();
+  }, [initializeSound]);
+
+  const { advanceStep, setVisible, step } = useTutorial();
+  const { setObserverId, clearObserverId } = useTutorialStore();
+  const [tutorialObserver, setTutorialObserver] = useState<null | number>(null);
+  useEffect(() => {
+    if (tutorialObserver !== null) {
+      unregisterObserver(tutorialObserver);
+      clearObserverId();
+    }
+    const observerId = registerObserver(
+      new TutorialObserver(advanceStep, setVisible, step),
+    );
+    setTutorialObserver(observerId);
+    setObserverId(observerId);
+  }, [advanceStep, setVisible, step]);
+
+  useEffect(() => {
+    setSoundDependency(playSoundEffect);
+    initializeAchievements();
+  }, [playSoundEffect, setSoundDependency, initializeAchievements]);
 
   return <RootNavigator />;
 }
