+++ frontend/package.json
@@ -66,7 +66,8 @@
     "react-native-webview": "13.13.5",
     "starknet": "^7.4.0",
     "tailwindcss": "^3.4.17",
-    "uuid": "^11.1.0"
+    "uuid": "^11.1.0",
+    "zustand": "^5.0.5"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
